Migrate to DSharpPlus 4.2.0-nightly-01105
This is a breaking change to all plugins!

On March 31st, Discord made a breaking API change with member objects, causing Json deserialization to throw an exception for DiscordMember. The lead developer of DCS is currently incapacitated, though D#+ has patched this issue by adding JsonIgnore to properties in DiscordMember which access base DiscordUser values.

This migration requires a change of all namespace imports, and every plugin will now have to be patched to account for this. Additionally, DCS offered a UsernameWithDiscriminator property - this has now been introduced as an extension method to accomplish parity.

This change will prolong development as the codebases will need to be brought up to speed on the new NuGet references.

This change does not require EF migration.